feat, fix
feat: + button clicke MainVIew Upload
Fix : add Model, modelStore
