--- youAndIConnection/youAndIConnection/Model/Resources/MemoryData.json
@@ -1,8 +1,0 @@
-[
-    {
-        "id": 1,
-        "title": "제목",
-        "description": "하는 할 수 있다."
-        
-    }
-]
